+++ utils/powershell/execute.js
@@ -18,7 +18,7 @@
  * The submodule of powershell
  * @module utils_powershell_execute
  */
-module.exports = async (...args) => {
+module.exports = (...args) => {
   return new Promise(async (resolve, reject) => {
     const [params, obj] = args;
     const [pathname, curdir] = params;